fix(compiler): suppress C4267 warning by explicit cast in register coalescing
【动机】
在寄存器分配的并查集合并过程中，`newAdj.size()` 返回 `size_t`，被隐式转换为 `int` 赋值给 `degree[i]`，触发 MSVC C4267 警告（从 size_t 转换到 int 可能丢失数据）。在 `/WX` 模式下，该警告被视为编译错误，阻塞构建。

【原理】
使用 `static_cast<int>(newAdj.size())` 显式转换，明确告知编译器转换是有意为之，消除警告。由于当前寄存器数量远小于 int 最大值，转换安全。

+++ src/regvm/compiler/RegisterAllocator.cpp
@@ -385,7 +385,7 @@ void RegisterAllocator::allocate(IRGraph* graph) {
                     if (rootN != i) newAdj.insert(rootN);
                 }
                 adj[i] = newAdj;
-                degree[i] = newAdj.size();
+                degree[i] = static_cast<int>(newAdj.size());
             }
         }
     }
